Bunch of changes for Win32. - The .def files are no longer used for exporting symbols. Instead, there is the PUBLIC_API macro. - Did some juggling with the command line arguments, linking libdeng2 temporarily to libdeng. - Fixed the complete CMake build. Installing into a runtime dir works, too ("vcbuild INSTALL.vcproj").

+++ doomsday/libdeng/portable/src/r_data.c
@@ -232,7 +232,7 @@ boolean R_SetDefaultColorPalette(colorpaletteid_t id)
  * @return              Color palette id.
  */
 colorpaletteid_t R_CreateColorPalette(const char* fmt, const char* name,
-                                      const byte* data, ushort num)
+                                      const byte* data, size_t num)
 {
 #define MAX_BPC         (16) // Max bits per component.
